backend/requirements.txt
modal==0.62.204
fastapi==0.111.0
pydantic==2.7.3
httpx==0.27.0
uvicorn[standard]==0.25.0
python-slugify==8.0.4
python-dotenv==1.0.1
python-multipart==0.0.9